Issue #13: Period in place of <BLANKLINE>
This is a slight hack, using a global BLANKLINE_MATCH string which can
be used to replace either `<BLANKLINE>` or `.` with a real blank line.
The reverse operation (replacing actual blank lines for reporting
purposes in `output_difference`) is left alone.

Also added more tests to ensure that the DONT_ACCEPT_BLANKLINE flag
still works as expected.