Start adding unit tests
It appears that some of the critical core methods in the doctest module
are not tested (either by doctests, or by any of the tests in the
official cpython repository). Because I'm likely to be mucking about
with these, I decided to start setting up some basic tests of their
behavior to detect breakage.