Change ellipsis from (...) to <...>
Angle-brackets are less common than parentheses in valid python code,
particularly when they are paired. Using `<...>` instead of `(...)` for
auto-ellipsis is more in keeping with the use of angle-brackets in
the existing `<BLANKLINE>` marker, and also looks slightly nicer when
used inline, e.g.:

    [2, 4, 6, <...>, 18, 20]

Also, move .rst files into an `rst` subdirectory, and include a unit
test for whitespace characters matching a `<BLANKLINE>`.